Add nozzle to player
It is implemented as a texture object, which means new exceptions for
collisions needed to be added. The one bug I've observed so far is when
colliding with an enemy, the nozzle tends to flicker for a fraction of a
second. Right now, the nozzle doesn't rotate either, it just points in
the direction the player was moved