Add functionality to add points from projectile collision back to player
There are also some minor tweaks to make the code more readable and
concise, as well as a partial mitigation for the segfaults that are
occurring with the collision between projectiles and enemies in certain
conditions (namely, firing consecutive times at an enemy). For testing
purposes, this was even reproduced with one enemy.