Fix segfaulting when more than 1 enemy dies at once
This was due to a slightly faulty method to remove dead objects. This
time, I created a function to find the predecessor to a dead object
and used that only if the object is dead. This also helps demystify that
part of the code as well. README.md has been updated accordingly

Besides that, there's some slight changes in the positions where
enemies render.