Add a nicer pause menu
This pretty much just prints some text right now, I plan to make it a
bit more elegant. In fact I was thinking of implementing a similar
feature to Panzer where the text function itself would set the
coordinates for the text by itself based on font size etc.

The text related functions were cleaned up as well, so now createText()
creates and renders text at once, which eliminates the need for the
custom textObj struct. Some minor cleanups here and there were also
done.

Right now the text is all green since it's easier to tell if green text
is showing up where it's not supposed to be (it's not!), but this is an
easy fix.